Validate availability in Board instead of separate class
