+++ spec/board_spec.rb
@@ -15,7 +15,7 @@ describe 'Board' do
     end
   end
 
-  context '.full?' do
+  context '#full?' do
     it 'returns true when the board is full' do
       board = Board.new
 
@@ -27,9 +27,25 @@ describe 'Board' do
     it 'returns false when the board is not full' do
       board = Board.new
 
-      board.values = ['X', 1, 'O', 'X', 'X', 'O', 'X', 'X', 'O']
+      board.values = ['X', 2, 'O', 'X', 'X', 'O', 'X', 'X', 'O']
 
       expect(board.full?).to eq(false)
+    end
+  end
+
+  context '#available?' do
+    it 'returns true when the space is available' do
+      board = Board.new
+      board.values = ['X', 2, 3, 'O', 5, 6, 'X', 8, 9]
+
+      expect(board.available?('6')).to eq(true)
+    end
+
+    it 'returns false when the space is not available' do
+      board = Board.new
+      board.values = ['X', 2, 3, 'O', 5, 6, 'X', 8, 9]
+
+      expect(board.available?('7')).to eq(false)
     end
   end
 end
