+++ spec/prompt_spec.rb
@@ -16,21 +16,21 @@ end
 class TestNumberValidator
 end
 
-class TestAvailabilityValidator
+class TestBoard
 end
 
 describe 'Prompt' do
   context '#call' do
     it 'returns valid user input' do
       console = TestConsole.new
-      availability_validator = TestAvailabilityValidator.new
+      board = TestBoard.new
       number_validator = TestNumberValidator.new
 
       allow(console).to receive(:gets_input).and_return(8)
       allow(number_validator).to receive(:valid?).and_return(true)
-      allow(availability_validator).to receive(:available?).and_return(true)
+      allow(board).to receive(:available?).and_return(true)
 
-      prompt = Prompt.new(console:, number_validator:, availability_validator:)
+      prompt = Prompt.new(console:, number_validator:, board:)
 
       message = 'Choose a space. Enter 1-9: '
       error_message = 'Please enter a valid number.'
@@ -41,14 +41,14 @@ describe 'Prompt' do
 
     it 'continues prompting the user until valid input is received' do
       console = TestConsole.new
-      availability_validator = TestAvailabilityValidator.new
+      board = TestBoard.new
       number_validator = TestNumberValidator.new
 
       allow(console).to receive(:gets_input).and_return('asdf', 1)
       allow(number_validator).to receive(:valid?).and_return(false, true)
-      allow(availability_validator).to receive(:available?).and_return(true)
+      allow(board).to receive(:available?).and_return(true)
 
-      prompt = Prompt.new(console:, number_validator:, availability_validator:)
+      prompt = Prompt.new(console:, number_validator:, board:)
 
       message = 'Choose a space. Enter 1-9: '
       error_message = 'Please enter a valid number.'
